Add files via upload
This update consists of some patches and fix ups of the older code. This will be the last update as I am going to be doing a rewrite of the code making it more efficient. This update adds support for installing dependencies for building Cemu *Only arch based distros are supported as it uses pacman to sync* other than that, the rest contains some refinements and the introduction of Path Manager for managing the paths to cemu and the mlc01 *currently still a wip but more progress has been made*