Add files via upload
This is a wip update, Arch is still the only supported distro but MacOS support is now being added as well as other linux distros. In this update you will find the new OS-Type.sh, this is now the preferred way to start CM as it will started according to the OS you choose, this is very important for MacOS users as the file system works differently. MacOS support is still in its early stages and will require you to first build Cemu yourself as CM is not able to download or build it yet. In the upcoming versions I will be adding an auto updater as well as MacOS build support. Files no longer needed - PathSet.sh